fts: extract reader/search.rs (scored search entry points)
Phase 2 (cont). Split the scored search entry points — search /
search_with_floor, the pretokenized and doc-id-ranged variants,
search_multi, the single-term BlockMaxWAND path, and the phrase-aware
atom scored walk — out of the core impl into reader/search.rs as its own
impl FtsReader block. The public search methods stay pub; widened the
ClauseLists predicate helpers, top_k, and OrCursorSet's fields to
pub(super). Pure move; FTS + reader + supertable-query suites green.
